requirements.txt
Django==3.2.12
django-cors-headers==3.11.0
django-csp==3.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytest==7.0.1
pytest-django==4.5.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
drf-yasg==1.20.0